piexchange/requirements.txt
Django
gunicorn
psycopg2-binary
PyYAML
djangorestframework
docker-py
dockerpty
docopt
requests
six
texttable
websocket-client
django_forms_bootstrap
dj_static
django-oauth-toolkit
django-cors-headers
django-allauth
django-compressor
httpie
